Refactor image loading functionality into Pixel class
